Refactor EntityCache to an interface
